requirements.txt
Flask
Flask-Session
requests
flask-migrate
flask-sqlalchemy
authlib
python-dotenv
requests
Flask-Mail
